Ignore google forms

+++ src/main/java/ua/kharkiv/elsol/exporter/service/ExporterRunner.java
@@ -42,26 +42,32 @@ public class ExporterRunner implements CommandLineRunner {
     List<File> allFiles = googleDriveService.getAllFiles(latestExportRun);
     AtomicInteger successfulNumberOfFiles = new AtomicInteger();
     AtomicInteger failedNumberOfFiles = new AtomicInteger();
+    AtomicInteger skippedFiles = new AtomicInteger();
     allFiles.forEach(file -> {
       LOGGER.info("file name: {}, mimeType: {}, modifiedTime: {}",
           file.getName(), file.getMimeType(), file.getModifiedTime());
       try {
         var filePath = googleDriveService.downloadFile(file);
-        LOGGER.info("File {}, was downloaded to: {}", file.getName(), filePath);
-        String folderPath = getFolderPath(file);
-        uploadService.uploadFile(filePath, folderPath);
-        Thread.sleep(3000);
-        successfulNumberOfFiles.incrementAndGet();
-        Files.delete(Path.of(filePath));
-        LOGGER.info("File {} was cleaned up.", filePath);
+        if (filePath != null) {
+          LOGGER.info("File {}, was downloaded to: {}", file.getName(), filePath);
+          String folderPath = getFolderPath(file);
+          uploadService.uploadFile(filePath, folderPath);
+          Thread.sleep(3000);
+          successfulNumberOfFiles.incrementAndGet();
+          Files.delete(Path.of(filePath));
+          LOGGER.info("File {} was cleaned up.", filePath);
+        } else {
+          skippedFiles.incrementAndGet();
+        }
       } catch (Exception ex) {
         LOGGER.warn("Unable to export file {}, reason: {}", file.getName(), ex);
         failedNumberOfFiles.incrementAndGet();
       }
     });
     String notificationMessage = MessageFormat.format(
-        "Export of google drive has finished. Total number of files: {0}, successfully exported: {1}, failed to export: {2}",
-        allFiles.size(), successfulNumberOfFiles.get(), failedNumberOfFiles.get());
+        "Export of google drive has finished. Total number of files: {0}, successfully exported: {1}, failed to export: {2}, skipped: {3}",
+        allFiles.size(), successfulNumberOfFiles.get(), failedNumberOfFiles.get(),
+        skippedFiles.get());
     notificationService.sendNotification(notificationMessage);
     dataStoreService.updateLatestExportRun(currentExportRunTime);
     LOGGER.info("Finish file exporter with the result: {}", notificationMessage);
